+++ CHANGELOG.rst
@@ -8,7 +8,10 @@ The format is based on `Keep a Changelog <https://keepachangelog.com>`_.
 
 2.0.0
 -----
-- The curve fitting parameters (top, bottom, slope) can now be constrained to a range in addition to being completely free or fixed. This means ``fixtop`` and ``fixbottom`` can be 2-tuples of bounds.
+- The curve fitting parameters (top, bottom, slope) can now be constrained to a range in addition to being completely free or fixed. Specifically:
+  - ``fixtop`` and ``fixbottom`` parameters to ``HillCurve`` can be 2-tuples of bounds
+  - addedd ``fixslope`` parameter to ``HillCurve``
+  - New ``constrain_params_range`` notebook tests and documents this functionality.
 
 - Add ``no_curve_fit_first`` argument to ``HillCurve`` to aid debugging/development.